pyproject.toml
[project]
name = "qbittorent-to-kindle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.2",
]
